+++ tiger-test-lib/src/main/java/de/gematik/test/tiger/lib/TigerDirector.java
@@ -31,6 +31,7 @@ import de.gematik.test.tiger.common.config.TigerGlobalConfiguration;
 import de.gematik.test.tiger.common.data.config.tigerproxy.TigerProxyConfiguration;
 import de.gematik.test.tiger.common.web.TigerBrowserUtil;
 import de.gematik.test.tiger.lib.exception.TigerStartupException;
+import de.gematik.test.tiger.lib.rbel.RbelMessageValidator;
 import de.gematik.test.tiger.lib.reports.TigerRestAssuredCurlLoggingFilter;
 import de.gematik.test.tiger.lib.serenityrest.SerenityRestUtils;
 import de.gematik.test.tiger.testenvmgr.TigerTestEnvMgr;
@@ -54,6 +55,7 @@ import java.util.concurrent.TimeUnit;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
 import net.serenitybdd.rest.SerenityRest;
 import org.apache.commons.io.IOUtils;
@@ -64,6 +66,7 @@ import org.springframework.boot.Banner.Mode;
 import org.springframework.boot.WebApplicationType;
 import org.springframework.boot.builder.SpringApplicationBuilder;
 import org.springframework.context.ConfigurableApplicationContext;
+import org.springframework.test.util.ReflectionTestUtils;
 
 /**
  * The TigerDirector is the public interface of the high level features of the Tiger test framework.
@@ -438,10 +441,14 @@ public class TigerDirector {
   }
 
   @VisibleForTesting
+  @SneakyThrows
   public static synchronized void testUninitialize() {
     initialized = false;
     tigerTestEnvMgr = null;
     curlLoggingFilter = null;
+
+    ReflectionTestUtils.setField(RbelMessageValidator.class, "instance", null);
+
     LocalProxyRbelMessageListener
         .clearTestingInstance(); // NOSONAR - the method testUninitialize should also only be used
     // for testing