+++ tiger-test-lib/src/main/java/de/gematik/test/tiger/lib/rbel/RbelMessageValidator.java
@@ -61,7 +61,6 @@ import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
 import lombok.val;
 import org.apache.commons.collections4.iterators.ReverseListIterator;
-import org.apache.commons.collections4.list.UnmodifiableList;
 import org.apache.commons.lang3.StringUtils;
 import org.awaitility.core.ConditionTimeoutException;
 import org.jetbrains.annotations.NotNull;
@@ -80,7 +79,7 @@ public class RbelMessageValidator {
 
   private static final List<String> EMPTY_PATH = List.of("", "/");
   public static final TigerTypedConfigurationKey<Integer> RBEL_REQUEST_TIMEOUT =
-      new TigerTypedConfigurationKey<>("tiger.rbel.request.timeout", Integer.class);
+      new TigerTypedConfigurationKey<>("tiger.rbel.request.timeout", Integer.class, 5);
 
   private static final Map<String, UnaryOperator<DiffBuilder>> DIFF_OPTIONS = new HashMap<>();
 
@@ -91,16 +90,31 @@ public class RbelMessageValidator {
     DIFF_OPTIONS.put("txtnormalize", DiffBuilder::normalizeWhitespace);
   }
 
-  private static final class InstanceHolder {
-    private static final RbelMessageValidator instance = new RbelMessageValidator();
-  }
+  private static RbelMessageValidator instance;
 
   public static RbelMessageValidator getInstance() {
-    return InstanceHolder.instance;
+    initializeInstance();
+    instance.registerJexlToolbox();
+    return instance;
+  }
+
+  private static synchronized void initializeInstance() {
+    // In unit tests when we reset the tiger test environment ( see
+    // de.gematik.test.tiger.lib.TigerDirector.testUninitialize )
+    // we set the instance to null, to force the recreation of the RbelMessageValidator. Otherwise
+    // the instance will keep references
+    // to a discarded tigerTestEnvMgr and tigerProxy and not see the messages of the current
+    // environment.
+    if (instance == null) {
+      instance = new RbelMessageValidator();
+    }
   }
 
   private final TigerTestEnvMgr tigerTestEnvMgr;
   private final TigerProxy tigerProxy;
+
+  @Getter(AccessLevel.PRIVATE)
+  private final LocalProxyRbelMessageListener localProxyRbelMessageListener;
 
   @Setter @Getter protected RbelElement currentRequest;
 
@@ -111,27 +125,35 @@ public class RbelMessageValidator {
   @Getter
   protected RbelElement currentResponse;
 
-  public RbelMessageValidator() {
+  private RbelMessageValidator() {
     this(
         TigerDirector.getTigerTestEnvMgr(),
-        TigerDirector.getTigerTestEnvMgr().getLocalTigerProxyOrFail());
+        TigerDirector.getTigerTestEnvMgr().getLocalTigerProxyOrFail(),
+        LocalProxyRbelMessageListener.getInstance());
   }
 
   @VisibleForTesting
-  public RbelMessageValidator(TigerTestEnvMgr tigerTestEnvMgr, TigerProxy tigerProxy) {
-    TigerJexlExecutor.registerAdditionalNamespace("rbel", new JexlToolbox());
+  public RbelMessageValidator(
+      TigerTestEnvMgr tigerTestEnvMgr,
+      TigerProxy tigerProxy,
+      LocalProxyRbelMessageListener localProxyRbelMessageListener) {
+    registerJexlToolbox();
     this.tigerTestEnvMgr = tigerTestEnvMgr;
     this.tigerProxy = tigerProxy;
+    this.localProxyRbelMessageListener = localProxyRbelMessageListener;
+  }
+
+  private void registerJexlToolbox() {
+    TigerJexlExecutor.registerAdditionalNamespace("rbel", new JexlToolbox());
   }
 
   public List<RbelElement> getRbelMessages() {
     tigerProxy.waitForAllCurrentMessagesToBeParsed();
-    return new UnmodifiableList<>(
-        new ArrayList<>(LocalProxyRbelMessageListener.getInstance().getValidatableRbelMessages()));
+    return localProxyRbelMessageListener.getValidatableRbelMessages().stream().toList();
   }
 
   public void clearRbelMessages() {
-    LocalProxyRbelMessageListener.getInstance().clearValidatableRbelMessages();
+    localProxyRbelMessageListener.clearValidatableRbelMessages();
   }
 
   @SuppressWarnings("java:S1135")
@@ -175,7 +197,7 @@ public class RbelMessageValidator {
         .filter(RbelRequestFacet::isResponseRequired)
         .isPresent()) {
       try {
-        final int requestTimeout = RBEL_REQUEST_TIMEOUT.getValue().orElse(5);
+        final int requestTimeout = RBEL_REQUEST_TIMEOUT.getValueOrDefault();
         await("Waiting for matching request")
             .atMost(requestTimeout, TimeUnit.SECONDS)
             .pollInterval(500, TimeUnit.MILLISECONDS)
@@ -244,7 +266,7 @@ public class RbelMessageValidator {
   }
 
   protected RbelElement findMessageByDescription(final RequestParameter requestParameter) {
-    final int waitsec = RBEL_REQUEST_TIMEOUT.getValue().orElse(5);
+    final int waitsec = RBEL_REQUEST_TIMEOUT.getValueOrDefault();
 
     Optional<RbelElement> initialElement = getInitialElement(requestParameter);
 
@@ -293,8 +315,7 @@ public class RbelMessageValidator {
   }
 
   private Optional<RbelElement> getInitialElement(RequestParameter requestParameter) {
-    var validatableRbelMessages =
-        LocalProxyRbelMessageListener.getInstance().getValidatableRbelMessages();
+    var validatableRbelMessages = localProxyRbelMessageListener.getValidatableRbelMessages();
     if (requestParameter.isStartFromLastMessage()) {
       var markerMessage =
           requestParameter.isRequireRequestMessage() ? currentRequest : lastFoundMessage;
@@ -742,7 +763,7 @@ public class RbelMessageValidator {
 
   public void readTgrFile(String filePath) {
     List<RbelElement> readElements = tigerProxy.readTrafficFromTgrFile(filePath);
-    readElements.forEach(LocalProxyRbelMessageListener.getInstance()::triggerNewReceivedMessage);
+    readElements.forEach(localProxyRbelMessageListener::triggerNewReceivedMessage);
   }
 
   public class JexlToolbox {
@@ -790,9 +811,7 @@ public class RbelMessageValidator {
 
     private RbelElement lastMessageMatching(Predicate<RbelElement> testMessage) {
       final Iterator<RbelElement> backwardsIterator =
-          LocalProxyRbelMessageListener.getInstance()
-              .getValidatableRbelMessages()
-              .descendingIterator();
+          localProxyRbelMessageListener.getValidatableRbelMessages().descendingIterator();
       while (backwardsIterator.hasNext()) {
         final RbelElement element = backwardsIterator.next();
         if (testMessage.test(element)) {