+++ tiger-test-lib/src/test/java/de/gematik/test/tiger/lib/rbel/AbstractRbelMessageValidatorTest.java
@@ -23,23 +23,17 @@ import de.gematik.rbellogger.RbelLogger;
 import de.gematik.rbellogger.captures.RbelFileReaderCapturer;
 import de.gematik.rbellogger.configuration.RbelConfiguration;
 import de.gematik.rbellogger.converter.initializers.RbelKeyFolderInitializer;
-import de.gematik.rbellogger.data.RbelElement;
-import de.gematik.rbellogger.util.IRbelMessageListener;
-import de.gematik.rbellogger.util.RbelMessagesSupplier;
-import de.gematik.test.tiger.LocalProxyRbelMessageListener;
 import de.gematik.test.tiger.common.config.TigerGlobalConfiguration;
 import de.gematik.test.tiger.glue.RBelValidatorGlue;
 import de.gematik.test.tiger.proxy.TigerProxy;
 import de.gematik.test.tiger.testenvmgr.TigerTestEnvMgr;
-import java.util.ArrayDeque;
-import java.util.Deque;
 import java.util.List;
 
 abstract class AbstractRbelMessageValidatorTest {
-  protected final TigerProxy tigerProxy = mock(TigerProxy.class);
-  protected final Deque<RbelElement> validatableMessagesMock = new ArrayDeque<>();
   protected RbelMessageValidator rbelMessageValidator;
   protected RBelValidatorGlue glue;
+  protected LocalProxyRbelMessageListenerTestAdapter localProxyRbelMessageListenerTestAdapter;
+  protected TigerProxy tigerProxy;
 
   protected void readTgrFileAndStoreForRbelMessageValidator(String rbelFile) {
     readTgrFileAndStoreForRbelMessageValidator(rbelFile, List.of());
@@ -54,39 +48,25 @@ abstract class AbstractRbelMessageValidatorTest {
                 .addInitializer(new RbelKeyFolderInitializer("src/test/resources"))
                 .addCapturer(RbelFileReaderCapturer.builder().rbelFile(rbelFile).build()));
     rbelLogger.getRbelCapturer().initialize();
-    validatableMessagesMock.addAll(rbelLogger.getMessageHistory());
+    localProxyRbelMessageListenerTestAdapter.addMessages(rbelLogger.getMessageHistory());
   }
 
   protected void setUp() {
     TigerGlobalConfiguration.reset();
-    validatableMessagesMock.clear();
-    LocalProxyRbelMessageListener.setTestingInstance(
-        new LocalProxyRbelMessageListener(
-            new RbelMessagesSupplier() {
-              @Override
-              public void addRbelMessageListener(IRbelMessageListener listener) {
-                // do nothing
-              }
 
-              @Override
-              public Deque<RbelElement> getRbelMessages() {
-                return validatableMessagesMock;
-              }
-            }));
+    this.localProxyRbelMessageListenerTestAdapter = new LocalProxyRbelMessageListenerTestAdapter();
 
-    when(tigerProxy.getRbelMessages()).thenReturn(validatableMessagesMock);
-    rbelMessageValidator = new RbelMessageValidator(mock(TigerTestEnvMgr.class), tigerProxy);
+    tigerProxy = mock(TigerProxy.class);
+    when(tigerProxy.getRbelMessages())
+        .thenReturn(localProxyRbelMessageListenerTestAdapter.getValidatableMessagesMock());
+
+    rbelMessageValidator =
+        new RbelMessageValidator(
+            mock(TigerTestEnvMgr.class),
+            tigerProxy,
+            localProxyRbelMessageListenerTestAdapter.getLocalProxyRbelMessageListener());
 
     rbelMessageValidator.clearCurrentMessages();
-
     glue = new RBelValidatorGlue(rbelMessageValidator);
-
-    LocalProxyRbelMessageListener.getInstance().clearValidatableRbelMessages();
-    RbelMessageValidator.RBEL_REQUEST_TIMEOUT.putValue(1);
-  }
-
-  protected static void tearDown() {
-    LocalProxyRbelMessageListener.clearTestingInstance();
-    RbelMessageValidator.RBEL_REQUEST_TIMEOUT.clearValue();
   }
 }